feat(variant): Route unsigned C++ integer constructors to UINT
`Variant(uint8_t)`, `Variant(uint16_t)`, `Variant(uint32_t)`,
`Variant(uint64_t)`, and `Variant(Math::uint_alt_t)` now select the
`Variant::UINT` carrier instead of storing a reinterpreted `int64_t`.

Boundaries that declare their Variant type elsewhere keep the signed
carrier, because nothing in the engine declares `UINT` yet:

- `VariantInternal::make()` retags unsigned results, so every bound
  method result, property value, and FoundryExtension conversion keeps
  the `Variant::INT` its own `GetTypeInfo<T>` declaration names.
- `MethodBind::set_default_arguments()` retags an unsigned default whose
  parameter declares `INT`.
- `BitField<T>` is an unsigned container for signed enum flags.
- `PropertyInfo::usage`, `MethodInfo::flags`, and `Connection::flags`
  are documented as `int` in script-facing reflection.
- The `ArrayMesh` `_surfaces` schema and the JSON native
  `PackedByteArray` element list are persisted signed-integer contracts.
- The debugger wire protocol declares its thread id field as a signed
  integer.

Reflection masks reaching script therefore still compose with `int`
operands, and a scene carrying an unsigned-typed native property saves,
reloads, and instantiates without resource corruption.

+++ core/debugger/remote_debugger.cpp
@@ -100,7 +100,9 @@ public:
 };
 
 Error RemoteDebugger::_put_msg(const String &p_message, const Array &p_data) {
-	Array msg = { p_message, Thread::get_caller_id(), p_data };
+	// The debugger wire protocol declares the thread id as a signed integer field, so the unsigned
+	// C++ `Thread::ID` is carried nominally.
+	Array msg = { p_message, int64_t(Thread::get_caller_id()), p_data };
 	Error err = peer->put_message(msg);
 	if (err != OK) {
 		n_messages_dropped++;
